[fix] Move i18n tests to separate config to avoid CI conflicts
