+++ scripts/i18n/collect-i18n-general.ts
@@ -1,11 +1,11 @@
 import * as fs from 'fs'
 
+import { comfyPageFixture as test } from '../../browser_tests/fixtures/ComfyPage'
 import { CORE_MENU_COMMANDS } from '../../src/constants/coreMenuCommands'
 import { SERVER_CONFIG_ITEMS } from '../../src/constants/serverConfig'
 import type { ComfyCommandImpl } from '../../src/stores/commandStore'
 import type { FormItem, SettingParams } from '../../src/types/settingTypes'
 import { formatCamelCase, normalizeI18nKey } from '../../src/utils/formatUtil'
-import { comfyPageFixture as test } from '../fixtures/ComfyPage'
 
 const localePath = './src/locales/en/main.json'
 const commandsPath = './src/locales/en/commands.json'
